+++ src/yafowil/widget/slider/__init__.py
@@ -26,7 +26,8 @@ resources.add(wr.ScriptResource(
 ))
 resources.add(wr.StyleResource(
     name='yafowil-slider-css',
-    resource='widget.css'
+    resource='widget.css',
+    compressed='widget.min.css'
 ))
 
 # B/C resources ##############################################################